==> 06_Map_imputed_data_to_coupon_data.ipynb <==
%config Complete.use_jedi = False

import pandas as pd
import numpy as np

%%time
starbucks_to_map = pd.read_csv("./starbucks_data/starbucks_df_per_coupon.csv")
starbucks_id_data = pd.read_csv("./starbucks_data/starbucks_df_per_id.csv")
starbucks_imputed = pd.read_csv("./starbucks_data/starbucks_imputed.csv")

starbucks_id_data.sort_values(["value", "gender", "age", "became_member_on"]).tail(2500).value.sum() == \
starbucks_imputed.sort_values(["value", "gender", "age", "became_member_on"]).tail(2500).value.sum()

starbucks_id_data.sort_values(["value", "gender", "age", "became_member_on"], inplace=True)
starbucks_imputed.sort_values(["value", "gender", "age", "became_member_on"], inplace=True)

starbucks_imputed["id"] = starbucks_id_data["id"]

starbucks_imputed = starbucks_imputed[["id", "gender", "income", "age"]]

starbucks_imputed

starbucks_to_map.drop(columns=["gender", "income", "age"], inplace=True)

starbucks_to_map.set_index("id", inplace=True)

starbucks_imputed.set_index("id", inplace=True)

starbucks_final = starbucks_to_map.join(starbucks_imputed, on="id", how="left")

starbucks_final.dtypes

starbucks_final.to_csv("./starbucks_data/starbucks_df_per_coupon_imputed.csv")
